Fix item spawing locations

+++ src/main/java/club/devcord/gamejam/utils/RelativeLocation.java
@@ -3,8 +3,8 @@ package club.devcord.gamejam.utils;
 import org.bukkit.Location;
 import org.bukkit.World;
 
-public record RelativeLocation(int x, int y, int z) {
-    public static RelativeLocation of(int x, int y, int z) {
+public record RelativeLocation(double x, double y, double z) {
+    public static RelativeLocation of(double x, double y, double z) {
         return new RelativeLocation(x, y, z);
     }
